+++ ports/Microchip/PIC16F-generic/api/pin/low.c.v
@@ -8,7 +8,7 @@ module pin
 
 // low puts a logic 0 to a pin
 @[inline]
-pub fn low(name int) {
+pub fn low(name u8) {
 	unsafe {
 		*(&C.PORTA + (name >> 3))  &=  ~(0x01 << (name - ((name >> 3) << 3)))
 	}// *(&C.PORTA + (name / 8))   &=  ~(0x01 << (name % 8))
